Added ability to define initializers using Lua, a very brief example of which can be found in Config/TypeOverloads.json, FrooxEngine.CommonToolPermissions
"Instance" is the current component that you're targeting, and any methods expected to be contained in it can be called.

Cleaned up overload handling slightly - it now only uses a single file for all types, and has less code duplication.

All paths now use BasePath as defined in the WikiAssetGenerator class. Redefine this to match your preferred output location, and copy the /Config/ directory to it whenever you pull a new version from the repository.