Fix for Issue #2: Removed NeosPluginManager from Launch Settings
Fix for Issue #3: Changed list type to INeosTypeProcessor as suggested

Fix for Issue #5: Solution is now configured to copy Neo.Lua.dll to Neos plugins folder on build, and now provides Neos.Lua.dll as a launch option to Neos in the launch settings. This will result in Neos loading the Neo.Lua binary into memory, but logging an error since it's not a valid Neos plugin - this however allows our actual plugin to access the types, without having to place the DLL in Managed_Data or in Neos' root.

Fix for Issue #6: Retargeted the project for .Net Framework 4.6.2 (oops!).

Additionally, moved project to VS2022, which moves launch settings to launchSettings.json